More cleanups and warning fixes

+++ src/sbs/sound.cpp
@@ -390,7 +390,7 @@ void Sound::SetPlayPosition(float percent)
 		unsigned int length;
 		sound->getLength(&length, FMOD_TIMEUNIT_MS);
 
-		unsigned int position = percent * length;
+		unsigned int position = (unsigned int)(percent * length);
 		channel->setPosition(position, FMOD_TIMEUNIT_MS);
 
 		position_queued = true;
